split async and sync
