+++ aiochclient/__init__.py
@@ -1,5 +1,6 @@
-from aiochclient.client import ChClient
-from aiochclient.exceptions import ChClientError
-from aiochclient.records import Record
+from aiochclient.async_db.client import AsyncClient
+from aiochclient.db.client import SyncClient
+from aiochclient.common.exceptions import ChClientError
+from aiochclient.common.records import Record
 
-__all__ = ["ChClient", "ChClientError", "Record"]
+__all__ = ["AsyncClient", "SyncClient", "ChClientError", "Record"]
