+++ aiochclient/common/exceptions.py
