+++ aiochclient/http_clients/aiohttp.py
@@ -2,7 +2,7 @@ from typing import Any, AsyncGenerator, List, Optional
 
 from aiohttp import ClientSession
 
-from aiochclient.exceptions import ChClientError
+from aiochclient.common.exceptions import ChClientError
 from aiochclient.http_clients.abc import HttpClientABC
