+++ requirements.txt
@@ -1,1 +1,8 @@
 sqlparse>=0.3.0
+ciso8601~=2.2.0
+httpx~=0.23.0
+aiohttp~=3.8.3
+setuptools~=65.5.0
+cython~=0.29.32
+pytest~=7.2.0
+trio~=0.22.0
